+++ src/info.js
@@ -1,0 +1,8 @@
+/**
+ *
+ *
+ * _____________________ Screeps AI _____________________
+ * @VIKING_repository: https://github.com/Jourloy/VIKING
+ * @author: JOURLOY
+ *
+ */